LDU Online Judge

+++ judge/stop.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-ps -e | grep judge | awk '{print "kill -9 " $1}' | sh
-ps -e | grep polling | awk '{print "kill -9 " $1}' | sh
-echo "polling: Stoped."
+polling_name=`ps -e | grep polling | awk '{print $4}'`
+if [ "${polling_name}" == "polling" ];then
+    ps -e | grep judge | awk '{print "kill -9 " $1}' | sh
+    ps -e | grep polling | awk '{print "kill -9 " $1}' | sh
+    echo -e "polling: Stoped."
+fi
